Added Weapons/Enemy Functionality
Added a pistol and ar to game, both are able to be grabbed. Functions are shooting, VFX blast, audio for gunshot, and physics update. Enemy is able to be shot and killed using these weapons. Weapons do different damage, weapons are able to shoot objects around the map. Objects also react when shot.